+++ test/augmenters/test_pooling.py
@@ -21,6 +21,7 @@ from imgaug import parameters as iap
 from imgaug.testutils import reseed
 
 
+# TODO add test that checks the padding behaviour
 class TestAveragePooling(unittest.TestCase):
     def setUp(self):
         reseed()
@@ -171,6 +172,7 @@ class TestAveragePooling(unittest.TestCase):
         assert params[0][1] is None
 
 
+# TODO add test that checks the padding behaviour
 # We don't have many tests here, because MaxPooling and AveragePooling derive
 # from the same base class, i.e. they share most of the methods, which are then
 # tested via TestAveragePooling.
@@ -215,6 +217,7 @@ class TestMaxPooling(unittest.TestCase):
         assert np.all(diff <= 1)
 
 
+# TODO add test that checks the padding behaviour
 # We don't have many tests here, because MinPooling and AveragePooling derive
 # from the same base class, i.e. they share most of the methods, which are then
 # tested via TestAveragePooling.
@@ -259,6 +262,7 @@ class TestMinPooling(unittest.TestCase):
         assert np.all(diff <= 1)
 
 
+# TODO add test that checks the padding behaviour
 # We don't have many tests here, because MedianPooling and AveragePooling
 # derive from the same base class, i.e. they share most of the methods, which
 # are then tested via TestAveragePooling.